bigfixes

+++ ml/train.py
@@ -104,8 +104,8 @@ def model(x, num_variables, fold, reuse=False):
         w2 = tf.get_variable('w2', shape=(hidden_nodes, 1), initializer=tf.random_normal_initializer())
         b2 = tf.get_variable('b2', shape=(1), initializer=tf.constant_initializer())
 
-    l1 = tf.tanh(tf.add(b1, tf.matmul(x, w1)))
-    logits = tf.add(b2, tf.matmul(l1, w2))
+    l1 = tf.tanh(tf.add(b1, tf.matmul(x, tf.cast(w1, tf.float64))))
+    logits = tf.add(b2, tf.matmul(l1, tf.cast(w2, tf.float64)))
     f = tf.nn.sigmoid(logits)
 
     return (w1, b1, w2, b2), f
